refactor WorkspaceQuickSearch
