+++ webui/react/src/pages/WorkspaceDetails/WorkspaceQuickSearchModalComponent.module.scss
